requirements.txt
asgiref==3.7.2
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
geographiclib==2.0
geopy==2.4.1
inflection==0.5.1
packaging==23.2
pillow==10.2.0
PyJWT==1.7.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
